Refactor mocks into multiple files
