--- front/src/pages/companies/CompaniesMockMode.tsx
@@ -1,10 +1,0 @@
-import { RecordTablePageMockMode } from '@/object-record/components/RecordTablePageMockMode';
-import { PageContainer } from '@/ui/layout/page/PageContainer';
-
-export const CompaniesMockMode = () => {
-  return (
-    <PageContainer>
-      <RecordTablePageMockMode />
-    </PageContainer>
-  );
-};
